Conflicts:
	Grammar.pp

+++ Visitor/Interpreter.php
@@ -319,7 +319,7 @@ class Interpreter implements \Hoa\Visitor\Visit {
                         return $value;
 
                     case 'number':
-                        return intval($value);
+                        return floatval($value);
 
                     case 'string':
                         return trim($value, '\'"');	//@todo attention ca trim trop!
